fix: Make lodash a dependency, rather than a dev dependency
I'm hoping this fixes my current error when using the
npm package.

Looking at the compiled outputs into js, the first line is:

import * as _ from "lodash";

But, because lodash is not a dependency, but rather a dev dependency,
this brings up an error when downloading and using from npm,
so the current npm package does not run

I am hoping that this simple fix will fix that.